requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.7
flasgger==0.9.7.1
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
validators==0.21.2
Werkzeug==2.3.7
